IMPL: tcl. axis_async_fifo.tcl now raises critical warnings, not errors
Previously, if there was an error while setting constraints for any FIFO
instance, the script would encounter an error and terminate, leading to
constraints not being applied for any subsequent FIFOs. Now, errors are
caught using a Tcl `catch` statement, and re-raised as critical
warnings, which do not cause the whole script to terminate.